Changes to project and tpk
Changed copyright headers. Created README for tpk sample. Finished tpk sample.
