+++ arcgis-ios-sdk-samples/Geometry/Project/ProjectViewController.swift
@@ -1,10 +1,16 @@
+// Copyright 2016 Esri.
 //
-//  ProjectViewController.swfit.swift
-//  ArcGIS Runtime SDK Samples
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
 //
-//  Created by Vivian Quach on 6/26/19.
-//  Copyright © 2019 Esri. All rights reserved.
+//   http://www.apache.org/licenses/LICENSE-2.0
 //
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 import UIKit
 import ArcGIS
